fix:  フィルタリング，マスク用のテストの修正する
